Made everything better
The x-ends have been modified to be stronger, the code is cleaner, anf
more general fixes.
nut_slot_hole is gone. nut_slot variable has been added to the nut_hole
function instead.
Took out the printed washers for the x_end_idler. use real ones instead,
it's easier and cleaner to print.
the belt clamp offset is set to be the thickness of the washer. it
should be set automatically, but I couldn't get the config file to look
all pretty that way (i know, what a shame).