Changed nut_hole to be stronger (star shaped/reinforced corners), increased tolerances for the bushing clamp and rod clamps.
The nut holes not have a small releif in the corners to reinforce them
and to allow the corners to be a bit cleaner. wel'll see how this works
out. use the screw hole test. I'm also changing all of the holes to use
the new vertical/horizontal code.
The tolerances for the bushing and smooth rod clamp holes have been
increased to allow for adjustments.